% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_binomial_name}
\alias{get_binomial_name}
\title{To extract binomial scientific names}
\usage{
get_binomial_name(v, patterns = "^([^ ]+ [^ ]*) .*", repl = "\\\\1", ...)
}
\arguments{
\item{v}{Input vector, must be a character vector.}

\item{patterns}{The Regex pattern to match, default will extract the first two words.}

\item{repl}{The values to replace the pattern matched that was specified by `patterns`.}

\item{...}{Additional arguments to be passed to `[sub]` function.}
}
\value{
A character vector, with the binomial scientific names
}
\description{
Some description about the function.
}
\examples{
get_binomial_name(c("Carex aquatilis Wahlenb.", "Boerhavia coulteri (Hook. f.) S. Watson"))

}
